several fixes for Apple silicon
cf https://sympa.inria.fr/sympa/arc/core-math/2026-01/msg00009.html

+++ src/binary64/erf/check_special.c
@@ -226,7 +226,7 @@ static void scan_consecutive(int64_t n, double x){
     n -= jmax;
     if( (count += jmax) > nreport){
       nreport += reportstride;
-      printf("n_tested = %ld x_current = %a\n", count, x);
+      printf("n_tested = %ld x_current = %a\n", (long) count, x);
     }
   }
 }
